9.0-agentic-rag/pyproject.toml
[project]
name = "9-0-agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langchain>=1.3.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.1",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.18",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.2.0",
    "pinecone[grpc]>=7.3.0",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
]
